+++ python/sedonadb/tests/test_funcs.py
@@ -15,6 +15,10 @@
 # specific language governing permissions and limitations
 # under the License.
 
+import pytest
+
+from sedonadb.expr.expression import AggregateUdf, ScalarUdf
+
 
 def test_random_geometry(con):
     df = con.funcs.table.sd_random_geometry("Point", 5, seed=99873)
@@ -24,3 +28,50 @@ def test_random_geometry(con):
 
     # Ensure the output is reproducible
     assert df.to_arrow_table() == df.to_arrow_table()
+
+
+def test_funcs_dir(con):
+    funcs_dir = dir(con.funcs)
+    assert "st_area" in funcs_dir
+    assert "st_buffer" in funcs_dir
+    assert "st_intersection" in funcs_dir
+
+
+def test_funcs_ipython_completions(con):
+    completions = con.funcs._ipython_key_completions_()
+    assert "st_area" in completions
+    assert "st_buffer" in completions
+
+
+def test_funcs_getitem_access(con):
+    st_area = con.funcs["st_area"]
+    assert isinstance(st_area, ScalarUdf)
+
+    e = st_area(con.col("geom"))
+    assert repr(e) == "Expr(st_area(geom))"
+
+    sum_fn = con.funcs["sum"]
+    assert isinstance(sum_fn, AggregateUdf)
+
+    e = sum_fn(con.col("x"))
+    assert repr(e) == "Expr(sum(x))"
+
+
+def test_funcs_getattr_not_found_raises_attribute_error(con):
+    with pytest.raises(AttributeError, match="Can't find scalar or aggregate function"):
+        con.funcs.nonexistent_function_xyz
+
+
+def test_funcs_getitem_not_found_raises_key_error(con):
+    with pytest.raises(KeyError, match="Can't find scalar or aggregate function"):
+        con.funcs["nonexistent_function_xyz"]
+
+
+def test_scalar_udf_repr(con):
+    st_area = con.funcs.st_area
+    assert repr(st_area) == "ScalarUdf(st_area)"
+
+
+def test_aggregate_udf_repr(con):
+    st_collect_agg = con.funcs.st_collect_agg
+    assert repr(st_collect_agg) == "AggregateUdf(st_collect_agg)"
